+++ app/src/main/java/de/baumann/browser/Browser_1.java
@@ -164,6 +164,29 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
         helper_main.grantPermissionsStorage(activity);
 
 
+        // show changelog
+
+        final String versionName = BuildConfig.VERSION_NAME;
+        String oldVersionName = sharedPref.getString("oldVersionName", "0.0");
+
+        if (!oldVersionName.equals(versionName)) {
+            AlertDialog.Builder builder = new AlertDialog.Builder(activity);
+            builder.setTitle(R.string.app_changelog);
+            builder.setMessage(helper_main.textSpannable(activity.getString(R.string.changelog_text)));
+            builder.setPositiveButton(
+                    getString(R.string.toast_yes),
+                    new DialogInterface.OnClickListener() {
+                        public void onClick(DialogInterface dialog, int id) {
+                            sharedPref.edit().putString("oldVersionName", versionName).apply();
+                            dialog.cancel();
+                        }
+                    });
+
+            AlertDialog alert = builder.create();
+            alert.show();
+        }
+
+
         // find Views
 
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
@@ -181,7 +204,7 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
 
         // setupViews
 
-        helper_browser.setupViews(activity, toolbar, mWebView, editText, imageButton, imageButton_left, imageButton_right, relativeLayout);
+        helper_browser.setupViews(activity, toolbar, mWebView, urlBar, editText, imageButton, imageButton_left, imageButton_right, relativeLayout);
         toolbar.setBackgroundColor(ContextCompat.getColor(activity, R.color.colorPrimary_1));
         setSupportActionBar(toolbar);
 
@@ -730,12 +753,6 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
         // automatically handle clicks on the Home/Up button, so long
         // as you specify a parent activity in AndroidManifest.xml.
         int id = item.getItemId();
-
-        if (id == R.id.action_search) {
-            urlBar.setVisibility(View.GONE);
-            editText.setVisibility(View.VISIBLE);
-            helper_editText.showKeyboard(activity, editText, 3, "", getString(R.string.app_search_hint));
-        }
 
         if (id == R.id.action_history) {
             helper_main.switchToActivity(activity, List_history.class, "", false);